ページ一覧画面にページタイトル，とそのリンクを張るようにした
Article.getArticleId() が Long を返すようになったので，
conf/routes, ArticleController も変更した

+++ app/controllers/ArticleController.java
@@ -12,7 +12,7 @@ public class ArticleController extends Controller {
     return ok(article_index.render(articles));
   }
 
-  public Result show(int id) {
+  public Result show(Long id) {
     return ok(show.render("show"));
   }
